ゲームプレイ中のUIのブーメランの残機の追加 ブーメランの数を３つに増やした ブーメランを拾えるように変更 警告文をなくした（シェーダー以外の）

+++ EmploymentWorks/Game/Object/Boomerang/State/BoomerangThrow.cpp
@@ -117,7 +117,7 @@ void BoomerangThrow::Enter()
 {
 	using namespace DirectX::SimpleMath;
 
-	m_index = m_spherePos.size() - 1;
+	m_index = static_cast<int> (m_spherePos.size() - 1);
 	m_startIndex = m_index;
 	m_transformRatio = 0;
 	m_totalTime = 0;
